+++ rusk-wallet/src/bin/command/tests.rs
@@ -29,7 +29,7 @@ async fn test_empty_history() {
 }
 
 #[tokio::test]
-async fn test_non_empty_history() {
+async fn test_history_transfer_convert() {
     configure_logger();
     wait_for_nodes_to_start().await.unwrap();
 
@@ -47,16 +47,11 @@ async fn test_non_empty_history() {
     )
     .await
     .unwrap();
-    // Need to wait for each transaction to be included in a block before moving
-    // to the next to ensure that the history ends up in the same order as the
-    // transactions made.
-    gql.wait_for(&rcv_moonlight).await.unwrap();
 
     let rcv_phoenix =
         rcv_phoenix_from_faucet(phoenix_addr.clone(), dusk(2500.0), gas_price)
             .await
             .unwrap();
-    gql.wait_for(&rcv_phoenix).await.unwrap();
 
     let mut txs_info = wait_for_tx_blocks_to_finalize(
         &gql,
@@ -74,6 +69,9 @@ async fn test_non_empty_history() {
     )
     .await
     .unwrap();
+    // Need to wait for each transaction to be included in a block before moving
+    // to the next to ensure that the history ends up in the same order as the
+    // transactions made.
     gql.wait_for(&moonlight_trans_to_other_wallet)
         .await
         .unwrap();
@@ -109,17 +107,6 @@ async fn test_non_empty_history() {
     .unwrap();
     gql.wait_for(&phoenix_to_moonlight).await.unwrap();
 
-    let stake_moonlight =
-        stake_moonlight(&mut wallet, &settings, dusk(1000.0).into(), gas_price)
-            .await
-            .unwrap();
-    gql.wait_for(&stake_moonlight).await.unwrap();
-
-    let stake_phoenix =
-        stake_phoenix(&mut wallet, &settings, dusk(1000.0).into(), gas_price)
-            .await
-            .unwrap();
-
     txs_info.extend(
         wait_for_tx_blocks_to_finalize(
             &gql,
@@ -128,8 +115,6 @@ async fn test_non_empty_history() {
                 &phoenix_trans_to_other_wallet,
                 &phoenix_to_moonlight,
                 &moonlight_to_phoenix,
-                &stake_moonlight,
-                &stake_phoenix,
             ],
         )
         .await
@@ -198,11 +183,105 @@ async fn test_non_empty_history() {
                 amount: 5_000.0,
                 fee: 0,
             },
+        ]
+    );
+}
+
+#[tokio::test]
+async fn test_history_stake_unstake() {
+    configure_logger();
+    wait_for_nodes_to_start().await.unwrap();
+
+    let (mut wallet, settings) = create_wallet().await.unwrap();
+    let gql = GraphQL::new(settings.state.clone(), status::headless).unwrap();
+    let moonlight_addr = wallet.default_address();
+    let phoenix_addr = wallet.shielded_account(0).unwrap();
+    let gas_price = 1;
+
+    let rcv_moonlight = rcv_moonlight_from_faucet(
+        moonlight_addr.clone(),
+        dusk(1500.0),
+        gas_price,
+    )
+    .await
+    .unwrap();
+
+    let rcv_phoenix =
+        rcv_phoenix_from_faucet(phoenix_addr.clone(), dusk(1500.0), gas_price)
+            .await
+            .unwrap();
+
+    let stake_moonlight =
+        stake_moonlight(&mut wallet, &settings, dusk(1000.0).into(), gas_price)
+            .await
+            .unwrap();
+    gql.wait_for(&stake_moonlight).await.unwrap();
+
+    let unstake_moonlight =
+        unstake_moonlight(&mut wallet, &settings, gas_price)
+            .await
+            .unwrap();
+    gql.wait_for(&unstake_moonlight).await.unwrap();
+
+    let stake_phoenix =
+        stake_phoenix(&mut wallet, &settings, dusk(1000.0).into(), gas_price)
+            .await
+            .unwrap();
+    gql.wait_for(&stake_phoenix).await.unwrap();
+
+    let unstake_phoenix = unstake_phoenix(&mut wallet, &settings, gas_price)
+        .await
+        .unwrap();
+
+    let txs_info = wait_for_tx_blocks_to_finalize(
+        &gql,
+        vec![
+            &rcv_moonlight,
+            &rcv_phoenix,
+            &stake_moonlight,
+            &unstake_moonlight,
+            &stake_phoenix,
+            &unstake_phoenix,
+        ],
+    )
+    .await
+    .unwrap();
+
+    let cmd = Command::History { profile_idx: None };
+    let history = cmd.run(&mut wallet, &settings).await.unwrap();
+    assert!(matches!(history, RunResult::History(_)));
+    let RunResult::History(tx_history) = history else {
+        unreachable!();
+    };
+    let tx_history: Vec<StrippedTxHistoryItem> =
+        tx_history.into_iter().map(|item| item.into()).collect();
+
+    assert_eq!(
+        tx_history,
+        vec![
+            // Receive money from faucet to moonlight address
+            StrippedTxHistoryItem {
+                direction: TransactionDirection::In,
+                amount: dusk(1500.0) as f64,
+                fee: txs_info[&rcv_moonlight].gas_spent * gas_price,
+            },
+            // Receive money from faucet to phoenix address
+            StrippedTxHistoryItem {
+                direction: TransactionDirection::In,
+                amount: dusk(1500.0) as f64,
+                fee: txs_info[&rcv_phoenix].gas_spent * gas_price,
+            },
             // Stake 1000 dusk with moonlight
             StrippedTxHistoryItem {
                 direction: TransactionDirection::Out,
                 amount: -(dusk(1000.0) as f64),
                 fee: txs_info[&stake_moonlight].gas_spent * gas_price,
+            },
+            // Unstake with moonlight
+            StrippedTxHistoryItem {
+                direction: TransactionDirection::Out,
+                amount: dusk(1000.0) as f64,
+                fee: txs_info[&unstake_moonlight].gas_spent * gas_price,
             },
             // Stake 1000 dusk with phoenix
             StrippedTxHistoryItem {
@@ -210,6 +289,12 @@ async fn test_non_empty_history() {
                 amount: -(dusk(1000.0) as f64),
                 fee: txs_info[&stake_phoenix].gas_spent * gas_price,
             },
+            // Unstake with phoenix
+            StrippedTxHistoryItem {
+                direction: TransactionDirection::Out,
+                amount: dusk(1000.0) as f64,
+                fee: txs_info[&unstake_phoenix].gas_spent * gas_price,
+            },
         ]
-    );
+    )
 }